+++ src/AccessIterator.php
@@ -1,0 +1,12 @@
+<?php
+namespace Pyncer\Iterable;
+
+use ArrayAccess;
+use Countable;
+use Iterator;
+use Pyncer\Iterable\AccessIteratorTrait;
+
+class AccessIterator implements ArrayAccess, Countable, Iterator
+{
+    use AccessIteratorTrait;
+}