+++ src/StaticAccessIterator.php
@@ -1,0 +1,12 @@
+<?php
+namespace Pyncer\Iterable;
+
+use ArrayAccess;
+use Countable;
+use Iterator;
+use Pyncer\Iterable\StaticAccessIteratorTrait;
+
+class StaticAccessIterator implements ArrayAccess, Countable, Iterator
+{
+    use StaticAccessIteratorTrait;
+}