+++ src/StaticSeekableIteratorTrait.php
@@ -1,21 +1,21 @@
 <?php
 namespace Pyncer\Iterable;
 
-use Pyncer\Iterable\IteratorTrait;
+use Pyncer\Iterable\StaticIteratorTrait;
 use Pyncer\Exception\OutOfBoundsException;
 
 use function array_key_exists;
 
-trait SeekableIteratorTrait
+trait StaticSeekableIteratorTrait
 {
-    use IteratorTrait;
+    use StaticIteratorTrait;
 
     public function seek(int $offset): void
     {
-        if (array_key_exists($offset, $this->keys)) {
+        if (array_key_exists($offset, static::$keys)) {
             throw new OutOfBoundsException('Invalid seek offset. (' . $offset . ')');
         }
 
-        $this->position = $offset;
+        static::$position = $offset;
     }
 }