+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Change Log
 
+## 1.1.0 - 2023-02-28
+
+### Added
+
+- Added Iterator class.
+- Added AccessIterator class.
+- Added AccessIteratorTrait trait.
+- Added SeekableIterator class.
+- Added StaticIteratorTrait trait.
+- Added StaticIterator class.
+- Added StaticAccessIteratorTrait trait.
+- Added StaticAccessIterator class.
+- Added StaticSeekableIteratorTrait trait.
+- Added StaticSeekableIterator class.
+
+### Fixed
+
+- SeekableIteratorTrait only working with list arrays.
+- Fixed missing serialization requirements on set and map interfaces.
+
 ## 1.0.0 - 2022-12-28
 
 Initial release.